TerminalCmd forward declaration
